// src/services/types/content-type-column-response.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const contentTypeColumnResponseFieldTypeEnum = {
	TEXT: "TEXT",
	BOOLEAN: "BOOLEAN",
	JSON: "JSON",
	NUMBER: "NUMBER",
	DATE: "DATE",
	ENUMERATION: "ENUMERATION",
	MEDIA: "MEDIA",
	RELATION: "RELATION",
} as const;

export type ContentTypeColumnResponseFieldTypeEnumKey =
	(typeof contentTypeColumnResponseFieldTypeEnum)[keyof typeof contentTypeColumnResponseFieldTypeEnum];

export const contentTypeColumnResponseTextTypeEnum = {SHORT_TEXT: "SHORT_TEXT", LONG_TEXT: "LONG_TEXT"} as const;

export type ContentTypeColumnResponseTextTypeEnumKey =
	(typeof contentTypeColumnResponseTextTypeEnum)[keyof typeof contentTypeColumnResponseTextTypeEnum];

export const contentTypeColumnResponseNumberTypeEnum = {
	INTEGER: "INTEGER",
	BIG_INTEGER: "BIG_INTEGER",
	DECIMAL: "DECIMAL",
	FLOAT: "FLOAT",
} as const;

export type ContentTypeColumnResponseNumberTypeEnumKey =
	(typeof contentTypeColumnResponseNumberTypeEnum)[keyof typeof contentTypeColumnResponseNumberTypeEnum];

export const contentTypeColumnResponseDateTypeEnum = {DATE: "DATE", DATE_TIME: "DATE_TIME", TIME: "TIME"} as const;

export type ContentTypeColumnResponseDateTypeEnumKey =
	(typeof contentTypeColumnResponseDateTypeEnum)[keyof typeof contentTypeColumnResponseDateTypeEnum];

export const contentTypeColumnResponseRelationTypeEnum = {
	ONE_TO_ONE: "ONE_TO_ONE",
	MANY_TO_ONE: "MANY_TO_ONE",
	ONE_TO_MANY: "ONE_TO_MANY",
	MANY_TO_MANY: "MANY_TO_MANY",
} as const;

export type ContentTypeColumnResponseRelationTypeEnumKey =
	(typeof contentTypeColumnResponseRelationTypeEnum)[keyof typeof contentTypeColumnResponseRelationTypeEnum];

export type ContentTypeColumnResponse = {
	/**
	 * @description 컬럼 명
	 * @type string
	 */
	name: string;
	/**
	 * @description 컬럼 순서
	 * @type integer, int32
	 */
	ordinalPosition: number;
	/**
	 * @description 컬럼 데이터 유형
	 * @type string
	 */
	columnType: string;
	/**
	 * @description 컬럼 필드 타입
	 * @type string
	 */
	fieldType: ContentTypeColumnResponseFieldTypeEnumKey;
	/**
	 * @description 필수값 여부
	 * @type boolean | undefined
	 */
	required?: boolean;
	/**
	 * @description 기본값
	 * @type string | undefined
	 */
	defaultValue?: string;
	/**
	 * @description 유니크 여부
	 * @type boolean | undefined
	 */
	unique?: boolean;
	/**
	 * @description 비공개 컬럼 여부
	 * @type boolean | undefined
	 */
	privateField?: boolean;
	/**
	 * @description 정규식 패턴
	 * @type string | undefined
	 */
	regexpPattern?: string;
	/**
	 * @description 최소 길이
	 * @type integer | undefined, int32
	 */
	minLength?: number;
	/**
	 * @description 최대 길이
	 * @type integer | undefined, int32
	 */
	maxLength?: number;
	/**
	 * @description 최소값
	 * @type integer | undefined, int32
	 */
	minValue?: number;
	/**
	 * @description 최대값
	 * @type integer | undefined, int32
	 */
	maxValue?: number;
	/**
	 * @description 텍스트 타입
	 * @type string | undefined
	 */
	textType?: ContentTypeColumnResponseTextTypeEnumKey;
	/**
	 * @description 숫자 타입
	 * @type string | undefined
	 */
	numberType?: ContentTypeColumnResponseNumberTypeEnumKey;
	/**
	 * @description 정밀도
	 * @type integer | undefined, int32
	 */
	precision?: number;
	/**
	 * @description 소수점 자릿수
	 * @type integer | undefined, int32
	 */
	scale?: number;
	/**
	 * @description 날짜 타입
	 * @type string | undefined
	 */
	dateType?: ContentTypeColumnResponseDateTypeEnumKey;
	/**
	 * @description enum 값 목록
	 * @type array | undefined
	 */
	enumValues?: string[];
	/**
	 * @description 관계 타입
	 * @type string | undefined
	 */
	relationType?: ContentTypeColumnResponseRelationTypeEnumKey;
	/**
	 * @description 관계 대상 테이블
	 * @type string | undefined
	 */
	relationTargetTable?: string;
};