// src/services/types/upsert-media-file-request.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

export const upsertMediaFileRequestProviderEnum = {AWS: "AWS", GCP: "GCP", NCP: "NCP", LOCAL: "LOCAL"} as const;

export type UpsertMediaFileRequestProviderEnumKey =
	(typeof upsertMediaFileRequestProviderEnum)[keyof typeof upsertMediaFileRequestProviderEnum];

export type UpsertMediaFileRequest = {
	/**
	 * @description 파일 URL
	 * @minLength 1
	 * @type string
	 */
	url: string;
	/**
	 * @description 스토리지 경로
	 * @minLength 1
	 * @type string
	 */
	path: string;
	/**
	 * @description MIME 타입
	 * @minLength 1
	 * @type string
	 */
	mimeType: string;
	/**
	 * @description 파일명
	 * @minLength 1
	 * @type string
	 */
	name: string;
	/**
	 * @description 파일 크기(byte)
	 * @minLength 1
	 * @type integer, int64
	 */
	size: number;
	/**
	 * @description 확장자
	 * @type string
	 */
	ext: string;
	/**
	 * @description 스토리지 제공자
	 * @type string
	 */
	provider: UpsertMediaFileRequestProviderEnumKey;
};